refactor: abstract useLoadWithRetry hook, use on profile loader
