+++ plugin-hrm-form/src/states/profile/hooks/useProfileLoader.ts
@@ -13,9 +13,10 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
-import { useCallback, useEffect, useRef } from 'react';
+import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import { useLoadWithRetry } from '../../hooks/useLoadWithRetry';
 import { RootState } from '../..';
 import asyncDispatch from '../../asyncDispatch';
 import * as ProfileActions from '../profiles';
@@ -34,7 +35,8 @@ type UseProfileLoaderReturn = {
  * Tools to load a profile by id into redux, by default it will load the profile automatically
  * @param {UseProfileLoaderParams} params - Parameters for the hook
  * @param params.profileId - The id of the profile to load
- * @param params.skipAutoload - If true, the profile will not be loaded automatically (default: false)
+ * @param params.autoload - If true, the profile will be loaded automatically (default: true)
+ * @param params.refresh - If changed to true, triggers a re-load (default: false)
  * @returns {UseProfileLoaderReturn} - loading state and actions for the profile
  */
 export const useProfileLoader = ({
@@ -52,42 +54,17 @@ export const useProfileLoader = ({
   }, [dispatch, profileId]);
 
   const firstFetch = autoload && !loading && !data && !error;
+  const safeToLoad = Boolean(profileId); // prevent load if there's no profile id
+  const shouldLoad = firstFetch || refresh; // load on initial mount
 
-  useEffect(() => {
-    if (!profileId) return;
-
-    if (firstFetch || refresh) {
-      loadProfile();
-    }
-  }, [firstFetch, loadProfile, profileId, refresh]);
-
-  const retryCount = useRef(0);
-  const backoff = useRef(1);
-  const retrying = useRef(false);
-  const timerId = useRef(null);
-
-  useEffect(() => {
-    // should retry
-    if (error && !loading && !retrying.current && retryCount.current < 10) {
-      // don't retry 4xx, the problem is on the client
-      if (error.status >= 400 && error.status < 500) {
-        return;
-      }
-
-      retrying.current = true;
-      timerId.current = setTimeout(() => {
-        loadProfile();
-        retrying.current = false;
-        retryCount.current += 1;
-        backoff.current *= 2;
-      }, 1000 * backoff.current);
-    }
-  }, [error, loadProfile, loading]);
-
-  // cleanup the retry timeout on unmount
-  useEffect(() => {
-    return () => clearTimeout(timerId.current);
-  }, []);
+  useLoadWithRetry({
+    error,
+    loadFunction: loadProfile,
+    loading,
+    retry: true,
+    safeToLoad,
+    shouldLoad,
+  });
 
   return {
     error,
